+++ python/paddle/distributed/fleet/meta_optimizers/localsgd_optimizer.py
@@ -13,6 +13,7 @@
 # limitations under the License.
 
 import paddle
+from paddle.incubate.nn.functional import autoincreased_step_counter
 from paddle.static import (
     default_main_program,
     default_startup_program,
@@ -113,7 +114,7 @@ class LocalSGDOptimizer(MetaOptimizerBase):
 
         p2s = self.create_snapshot_vars(main_block.program)
         with program_guard(main_block.program, startup_program):
-            step = paddle.fluid.layers.autoincreased_step_counter(begin=1)
+            step = autoincreased_step_counter(begin=1)
             k_steps = paddle.static.create_global_var(
                 name="k_steps",
                 shape=[1],
@@ -330,7 +331,7 @@ class AdaptiveLocalSGDOptimizer(MetaOptimizerBase):
 
         p2s = self.create_snapshot_vars(main_block.program)
         with program_guard(main_block.program, startup_program):
-            step = paddle.fluid.layers.autoincreased_step_counter(begin=1)
+            step = autoincreased_step_counter(begin=1)
 
             k_steps = paddle.static.create_global_var(
                 name="k_steps",
