+++ python/paddle/fluid/contrib/tests/test_multi_precision_fp16_train.py
@@ -293,7 +293,7 @@ class TestAmpWithNonIterableDataLoader(unittest.TestCase):
                 one_var = fluid.layers.fill_constant(
                     shape=[1], dtype='int64', value=1
                 )
-                with fluid.layers.control_flow.Switch() as switch:
+                with paddle.incubate.nn.Switch() as switch:
                     with switch.case(label != zero_var):
                         paddle.assign(zero_var, output=label)
                     with switch.default():
