+++ python/paddle/fluid/layers/learning_rate_scheduler.py
@@ -52,7 +52,7 @@ __all__ = [
 
 def _decay_step_counter(begin=0):
     # the first global step is zero in learning rate decay
-    global_step = nn.autoincreased_step_counter(
+    global_step = paddle.incubate.nn.functional.autoincreased_step_counter(
         counter_name='@LR_DECAY_COUNTER@', begin=begin, step=1
     )
     global_step = paddle.cast(global_step, 'float32')
@@ -359,7 +359,7 @@ def polynomial_decay(
                     shape=[1], dtype='float32', value=1.0
                 )
 
-                with control_flow.Switch() as switch:
+                with paddle.incubate.nn.Switch() as switch:
                     with switch.case(global_step == zero_var):
                         paddle.assign(one_var, output=div_res)
                 decay_steps = decay_steps * div_res
@@ -433,7 +433,7 @@ def piecewise_decay(boundaries, values):
                 name="learning_rate",
             )
 
-            with control_flow.Switch() as switch:
+            with paddle.incubate.nn.Switch() as switch:
                 for i in range(len(boundaries)):
                     boundary_val = tensor.fill_constant(
                         shape=[1],
@@ -590,7 +590,7 @@ def linear_lr_warmup(learning_rate, warmup_steps, start_lr, end_lr):
 
             global_step = _decay_step_counter()
 
-            with control_flow.Switch() as switch:
+            with paddle.incubate.nn.Switch() as switch:
                 with switch.case(global_step < warmup_steps):
                     decayed_lr = start_lr + linear_step * (
                         global_step / float(warmup_steps)
