+++ python/paddle/incubate/nn/functional/__init__.py
@@ -18,6 +18,7 @@ from .fused_transformer import fused_multi_transformer
 from .fused_matmul_bias import fused_matmul_bias, fused_linear
 from .fused_transformer import fused_bias_dropout_residual_layer_norm
 from .fused_ec_moe import fused_ec_moe
+from .autoincreased_step_counter import autoincreased_step_counter
 
 __all__ = [
     'fused_multi_head_attention',
@@ -27,4 +28,5 @@ __all__ = [
     'fused_linear',
     'fused_bias_dropout_residual_layer_norm',
     'fused_ec_moe',
+    'autoincreased_step_counter',
 ]
